requirements.txt
psutil
torchvision-nightly
fastai==1.0.28
